+++ src/datalevin/search.clj
@@ -282,6 +282,12 @@
 (declare doc-ref->id remove-doc* add-doc* hydrate-query display-xf score-docs
          get-rawtext new-search-engine)
 
+(def default-search-opts {:display             :refs
+                          :top                 10
+                          :proximity-expansion 2
+                          :proximity-max-dist  45
+                          :doc-filter          (constantly true)})
+
 (deftype SearchEngine [lmdb
                        analyzer
                        query-analyzer
@@ -296,7 +302,8 @@
                        ^AtomicInteger max-doc
                        ^AtomicInteger max-term
                        index-position?
-                       include-text?]
+                       include-text?
+                       search-opts]
   ISearchEngine
   (add-doc [this doc-ref doc-text check-exist?]
     (locking docs
@@ -330,11 +337,14 @@
     (.search this query {}))
   (search [this query {:keys [display top proximity-expansion
                               proximity-max-dist doc-filter]
-                       :or   {display             c/default-display
-                              top                 c/default-top
-                              proximity-expansion c/default-proximity-expansion
-                              proximity-max-dist  c/default-proximity-max-dist
-                              doc-filter          c/default-doc-filter}}]
+                       :or   {display (:display search-opts)
+                              top     (:top search-opts)
+                              proximity-expansion
+                              (:proximity-expansion search-opts)
+                              proximity-max-dist
+                              (:proximity-max-dist search-opts)
+                              doc-filter
+                              (:doc-filter search-opts)}}]
     (when-not (s/blank? query)
       (let [tokens (->> (query-analyzer query)
                         (mapv first)
@@ -763,14 +773,21 @@
     (l/visit lmdb docs-dbi load [:all-back])
     [@max-id norms docs]))
 
+(def default-opts {:analyzer        a/en-analyzer
+                   :index-position? false
+                   :include-text?   false
+                   :search-opts     default-search-opts})
+
 (defn new-search-engine
   ([lmdb]
    (new-search-engine lmdb nil))
-  ([lmdb {:keys [domain analyzer query-analyzer index-position? include-text?]
-          :or   {domain          "datalevin"
-                 analyzer        a/en-analyzer
-                 index-position? false
-                 include-text?   false}}]
+  ([lmdb {:keys [domain analyzer query-analyzer index-position? include-text?
+                 search-opts]
+          :or   {analyzer        (default-opts :analyzer)
+                 index-position? (default-opts :index-position?)
+                 include-text?   (default-opts :include-text?)
+                 search-opts     (default-opts :search-opts)
+                 domain          c/default-domain}}]
    (let [terms-dbi     (str domain "/" c/terms)
          docs-dbi      (str domain "/" c/docs)
          positions-dbi (str domain "/" c/positions)
@@ -792,7 +809,8 @@
                        (AtomicInteger. max-doc)
                        (AtomicInteger. max-term)
                        index-position?
-                       include-text?)))))
+                       include-text?
+                       search-opts)))))
 
 (defn transfer
   "transfer state of an existing engine to an new engine that has a
@@ -812,7 +830,8 @@
                   (.-max-doc old)
                   (.-max-term old)
                   (.-index-position? old)
-                  (.-include-text? old)))
+                  (.-include-text? old)
+                  (.-search-opts old)))
 
 (defprotocol IIndexWriter
   (write [this doc-ref doc-text])