+++ test/datalevin/core_test.clj
@@ -7,7 +7,7 @@
    [datalevin.test.core :as tdc :refer [db-fixture]]
    [clojure.test :as t :refer [is deftest testing use-fixtures]])
   (:import
-   [java.util UUID Arrays]
+   [java.util UUID Arrays Date]
    [java.lang Thread]))
 
 (use-fixtures :each db-fixture)
@@ -270,6 +270,13 @@
              @conn
              t50)
            #{[t100] [now]}))
+    (is (= 30 (sut/q '[:find ?n .
+                       :in $ ?t2
+                       :where
+                       [?e :foo/num ?n]
+                       [?e :foo/date ?t1]
+                       [(< ?t1 ?t2)]]
+                     @conn t50)))
     (sut/close conn)
     (u/delete-files dir)))